+++ src/GlobalPayments.Api/Entities/Enums/LodgingItemType.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace GlobalPayments.Api.Entities.Enums
+{
+    public enum LodgingItemType
+    {
+         RESTAURANT, 
+         GIFT_SHOP,
+         MINI_BAR,
+         PHONE,
+         LAUNDRY,
+         OTHER,
+         NO_SHOW
+    }
+}