+++ src/GlobalPayments.Api/Entities/Enums/PaymentMethodProgram.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace GlobalPayments.Api.Entities.Enums
+{
+    public enum PaymentMethodProgram
+    {
+        ASSURED_RESERVATION,
+        CARD_DEPOSIT,
+        PURCHASE,
+        OTHER
+    }
+}