+++ src/GlobalPayments.Api/Services/GpApiService.cs
@@ -6,16 +6,8 @@ using System.Text;
 
 namespace GlobalPayments.Api.Services {
     public class GpApiService {
-        public static AccessTokenInfo GenerateTransactionKey(Entities.Environment environment, string appId, string appKey, int? secondsToExpire = null, IntervalToExpire? intervalToExpire = null, string[] permissions = null) {
-            var connector = new GpApiConnector {
-                AppId = appId,
-                AppKey = appKey,
-                SecondsToExpire = secondsToExpire,
-                IntervalToExpire = intervalToExpire,
-                ServiceUrl = environment.Equals(Entities.Environment.PRODUCTION) ? ServiceEndpoints.GP_API_PRODUCTION : ServiceEndpoints.GP_API_TEST,
-                Timeout = 10000,
-                Permissions = permissions,
-            };
+        public static AccessTokenInfo GenerateTransactionKey(GpApiConfig gpApiConfig) {
+            GpApiConnector connector = new GpApiConnector(gpApiConfig);
 
             var data = connector.GetAccessToken();