+++ tests/GlobalPayments.Api.Tests/GpApi/GpApiAuthenticationTests.cs
@@ -7,7 +7,8 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 namespace GlobalPayments.Api.Tests.GpApi {
     [TestClass]
     public class GpApiAuthenticationTests : BaseGpApiTests {
-        CreditCardData card;
+        private CreditCardData card;
+        private GpApiConfig gpApiConfig;
         private const Environment ENVIRONMENT = Environment.TEST;
         private const string APP_ID = "JF2GQpeCrOivkBGsTRiqkpkdKp67Gxi0";
         private const string APP_KEY = "y7vALnUtFulORlTV";
@@ -20,20 +21,27 @@ namespace GlobalPayments.Api.Tests.GpApi {
                 ExpYear = expYear,
                 Cvn = "123",
             };
+
+            gpApiConfig = new GpApiConfig {
+                Environment = ENVIRONMENT, 
+                AppId = APP_ID,
+                AppKey = APP_KEY,
+                RequestLogger = new RequestConsoleLogger()
+            };
         }
 
         [TestMethod]
         public void GenerateAccessTokenManual() {
-            var info = GpApiService.GenerateTransactionKey(ENVIRONMENT, APP_ID, APP_KEY);
-            assertAccessTokenResponse(info);
+            var info = GpApiService.GenerateTransactionKey(gpApiConfig);
+            AssertAccessTokenResponse(info);
         }
 
         [TestMethod]
         public void GenerateAccessTokenManualWithPermissions() {
             var permissions = new[] { "PMT_POST_Create", "PMT_POST_Detokenize" };
-
-            var info =
-                GpApiService.GenerateTransactionKey(ENVIRONMENT, APP_ID, APP_KEY, permissions: permissions);
+            gpApiConfig.Permissions = permissions;
+            
+            var info = GpApiService.GenerateTransactionKey(gpApiConfig);
 
             Assert.IsNotNull(info);
             Assert.IsNotNull(info.Token);
@@ -46,55 +54,55 @@ namespace GlobalPayments.Api.Tests.GpApi {
         [TestMethod]
         public void GenerateAccessTokenManualWithWrongPermissions() {
             var permissions = new[] { "TEST_1", "TEST_2" };
-
+            gpApiConfig.Permissions = permissions;
+            
+            var exceptionCaught = false;
             try {
-                GpApiService.GenerateTransactionKey(ENVIRONMENT, APP_ID, APP_KEY, permissions: permissions);
+                GpApiService.GenerateTransactionKey(gpApiConfig);
             }
             catch (GatewayException ex) {
+                exceptionCaught = true;
                 Assert.AreEqual("INVALID_REQUEST_DATA", ex.ResponseCode);
                 Assert.AreEqual("40119", ex.ResponseMessage);
                 Assert.IsTrue(ex.Message.StartsWith("Status Code: BadRequest - Invalid permissions"));
+            } finally {
+                Assert.IsTrue(exceptionCaught);
             }
         }
 
         [TestMethod]
         public void ChargeCard_WithLimitedPermissions() {
             var permissions = new[] { "TRN_POST_Capture" };
-
-            ServicesContainer.ConfigureService(new GpApiConfig {
-                Environment = Environment.TEST,
-                AppId = "JF2GQpeCrOivkBGsTRiqkpkdKp67Gxi0",
-                AppKey = "y7vALnUtFulORlTV",
-                Permissions = permissions,
-                SecondsToExpire = 60
-            });
+            gpApiConfig.Permissions = permissions;
+            gpApiConfig.SecondsToExpire = 60;
+            GpApiService.GenerateTransactionKey(gpApiConfig);
+            
+            ServicesContainer.ConfigureService(gpApiConfig, "GpApiConfig_2");
 
             var exceptionCaught = false;
             try {
                 card.Charge()
                     .WithAmount(1.00m)
                     .WithCurrency("USD")
-                    .Execute();
+                    .Execute("GpApiConfig_2");
             }
             catch (GatewayException ex) {
                 exceptionCaught = true;
                 Assert.AreEqual("ACTION_NOT_AUTHORIZED", ex.ResponseCode);
                 Assert.AreEqual("40212", ex.ResponseMessage);
                 Assert.AreEqual("Status Code: Forbidden - Permission not enabled to execute action", ex.Message);
-            }
-            finally {
+            } finally {
                 Assert.IsTrue(exceptionCaught);
             }
         }
         
         [TestMethod]
         public void GenerateAccessTokenManual_AccessToken() {
-            var info =
-                GpApiService.GenerateTransactionKey(ENVIRONMENT, APP_ID, APP_KEY);
-            
-            ServicesContainer.ConfigureService(new GpApiConfig {
-                AccessTokenInfo = info
-            });
+            var info = GpApiService.GenerateTransactionKey(gpApiConfig);
+
+            gpApiConfig.AccessTokenInfo = info;
+
+            ServicesContainer.ConfigureService(gpApiConfig);
             
             var response = card.Verify()
                 .WithCurrency("USD")
@@ -107,13 +115,12 @@ namespace GlobalPayments.Api.Tests.GpApi {
 
         [TestMethod]
         public void GenerateAccessTokenManualWithSecondsToExpire() {
-            ServicesContainer.ConfigureService(new GpApiConfig {
-                Environment = Environment.TEST,
-                AppId = "JF2GQpeCrOivkBGsTRiqkpkdKp67Gxi0",
-                AppKey = "y7vALnUtFulORlTV",
-                SecondsToExpire = 60
-            });
+            gpApiConfig.SecondsToExpire = 60;
+            var info = GpApiService.GenerateTransactionKey(gpApiConfig);
+            gpApiConfig.AccessTokenInfo = info;
 
+            ServicesContainer.ConfigureService(gpApiConfig);
+            
             var response = card.Verify()
                 .WithCurrency("USD")
                 .Execute();
@@ -125,36 +132,46 @@ namespace GlobalPayments.Api.Tests.GpApi {
 
         [TestMethod]
         public void GenerateAccessTokenManualWithMaximumSecondsToExpire() {
+            var exceptionCaught = false;
             try {
-                GpApiService.GenerateTransactionKey(ENVIRONMENT, APP_ID, APP_KEY, 604801);
+                gpApiConfig.SecondsToExpire = 604801;
+                GpApiService.GenerateTransactionKey(gpApiConfig);
             }
             catch (GatewayException ex) {
+                exceptionCaught = true;
                 Assert.AreEqual("INVALID_REQUEST_DATA", ex.ResponseCode);
                 Assert.AreEqual("40213", ex.ResponseMessage);
                 Assert.AreEqual("Status Code: BadRequest - seconds_to_expire contains unexpected data", ex.Message);
+            } finally {
+                Assert.IsTrue(exceptionCaught);
             }
         }
         
         [TestMethod]
         public void GenerateAccessTokenManualWithInvalidSecondsToExpire() {
+            var exceptionCaught = false;
             try {
-                GpApiService.GenerateTransactionKey(ENVIRONMENT, APP_ID, APP_KEY, 10);
+                gpApiConfig.SecondsToExpire = 10;
+                GpApiService.GenerateTransactionKey(gpApiConfig);
             }
             catch (GatewayException ex) {
+                exceptionCaught = true;
                 Assert.AreEqual("INVALID_REQUEST_DATA", ex.ResponseCode);
                 Assert.AreEqual("40213", ex.ResponseMessage);
                 Assert.AreEqual("Status Code: BadRequest - seconds_to_expire contains unexpected data", ex.Message);
+            } finally {
+                Assert.IsTrue(exceptionCaught);
             }
         }
 
         [TestMethod]
         public void GenerateAccessTokenManualWithIntervalToExpire() {
-            ServicesContainer.ConfigureService(new GpApiConfig {
-                Environment = Environment.TEST,
-                AppId = "JF2GQpeCrOivkBGsTRiqkpkdKp67Gxi0",
-                AppKey = "y7vALnUtFulORlTV",
-                IntervalToExpire = IntervalToExpire.THREE_HOURS
-            });
+            gpApiConfig.IntervalToExpire = IntervalToExpire.THREE_HOURS;
+            var info = GpApiService.GenerateTransactionKey(gpApiConfig);
+
+            gpApiConfig.AccessTokenInfo = info;
+
+            ServicesContainer.ConfigureService(gpApiConfig);
 
             var response = card.Verify()
                 .WithCurrency("USD")
@@ -167,13 +184,9 @@ namespace GlobalPayments.Api.Tests.GpApi {
 
         [TestMethod]
         public void GenerateAccessTokenManualWithSecondsToExpireAndIntervalToExpire() {
-            ServicesContainer.ConfigureService(new GpApiConfig {
-                Environment = Environment.TEST,
-                AppId = "JF2GQpeCrOivkBGsTRiqkpkdKp67Gxi0",
-                AppKey = "y7vALnUtFulORlTV",
-                SecondsToExpire = 60,
-                IntervalToExpire = IntervalToExpire.FIVE_MINUTES
-            });
+            gpApiConfig.SecondsToExpire = 60;
+            gpApiConfig.IntervalToExpire = IntervalToExpire.THREE_HOURS;
+            GpApiService.GenerateTransactionKey(gpApiConfig);
 
             var response = card.Verify()
                 .WithCurrency("USD")
@@ -186,25 +199,35 @@ namespace GlobalPayments.Api.Tests.GpApi {
 
         [TestMethod]
         public void GenerateAccessTokenWithWrongAppId() {
+            var exceptionCaught = false;
             try {
-                GpApiService.GenerateTransactionKey(ENVIRONMENT, APP_ID + "a", APP_KEY);
+                gpApiConfig.AppId += "a";
+                GpApiService.GenerateTransactionKey(gpApiConfig);
             }
             catch (GatewayException ex) {
+                exceptionCaught = true;
                 Assert.AreEqual("ACTION_NOT_AUTHORIZED", ex.ResponseCode);
                 Assert.AreEqual("40004", ex.ResponseMessage);
                 Assert.AreEqual("Status Code: Forbidden - App credentials not recognized", ex.Message);
+            } finally {
+                Assert.IsTrue(exceptionCaught);
             }
         }
 
         [TestMethod]
         public void GenerateAccessTokenWithWrongAppKey() {
+            var exceptionCaught = false;
             try {
-                GpApiService.GenerateTransactionKey(ENVIRONMENT, APP_ID, APP_KEY + "a");
+                gpApiConfig.AppKey += "a";
+                GpApiService.GenerateTransactionKey(gpApiConfig);
             }
             catch (GatewayException ex) {
+                exceptionCaught = true;
                 Assert.AreEqual("ACTION_NOT_AUTHORIZED", ex.ResponseCode);
                 Assert.AreEqual("40004", ex.ResponseMessage);
                 Assert.AreEqual("Status Code: Forbidden - Credentials not recognized to create access token.", ex.Message);
+            } finally {
+                Assert.IsTrue(exceptionCaught);
             }
         }
 
@@ -220,15 +243,19 @@ namespace GlobalPayments.Api.Tests.GpApi {
                 }
             });
 
+            var exceptionCaught = false;
             try {
                 card.Verify()
                     .WithCurrency("USD")
                     .Execute();
             }
             catch (GatewayException ex) {
+                exceptionCaught = true;
                 Assert.AreEqual("NOT_AUTHENTICATED", ex.ResponseCode);
                 Assert.AreEqual("40001", ex.ResponseMessage);
                 Assert.AreEqual("Status Code: Unauthorized - Invalid access token", ex.Message);
+            } finally {
+                Assert.IsTrue(exceptionCaught);
             }
         }
 
@@ -244,19 +271,23 @@ namespace GlobalPayments.Api.Tests.GpApi {
                 }
             });
 
+            var exceptionCaught = false;
             try {
                 card.Verify()
                     .WithCurrency("USD")
                     .Execute();
             }
             catch (GatewayException ex) {
+                exceptionCaught = true;
                 Assert.AreEqual("NOT_AUTHENTICATED", ex.ResponseCode);
                 Assert.AreEqual("40001", ex.ResponseMessage);
                 Assert.AreEqual("Status Code: Unauthorized - Invalid access token", ex.Message);
+            } finally {
+                Assert.IsTrue(exceptionCaught);
             }
         }
 
-        private void assertAccessTokenResponse(AccessTokenInfo accessTokenInfo) {
+        private static void AssertAccessTokenResponse(AccessTokenInfo accessTokenInfo) {
             Assert.IsNotNull(accessTokenInfo);
             Assert.IsNotNull(accessTokenInfo.Token);
             Assert.AreEqual("Tokenization", accessTokenInfo.TokenizationAccountName);